improve styles
