+++ frontend/src/routes/login/+page.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+    import PageHeader from '$lib/components/PageHeader.svelte';
+    import Login from '$lib/components/LoginPage/Login.svelte';
+</script>
+
+<PageHeader title="Log in" subtitle="Log in with an OAuth provider of your choice." />
+<Login />
